Add missing .cpp files where only headers exist.
This uncovered several instances of non-self-contained header files
which are also fixed.

+++ Browser/CategoryPage.cpp
@@ -48,7 +48,7 @@ Browser::BrowserPage* Browser::CategoryPage::activateChild( const QModelIndex& i
     DB::ImageSearchInfo info = searchInfo();
 
     info.addAnd( m_category->name(), name );
-    if (static_cast<uint>(DB::ImageDB::instance()->search(info).size()) <= Settings::SettingsData::instance()->autoShowThumbnailView()) {
+    if (DB::ImageDB::instance()->search(info).size() <= Settings::SettingsData::instance()->autoShowThumbnailView()) {
         browser()->addAction( new Browser::OverviewPage( Breadcrumb(name), info, browser() ) );
         return new ImageViewPage( info, browser() );
     } else
